# api/openapi.yaml
openapi: 3.0.0
info:
  title: Mithlond Agent API
  version: 0.1.0
  license:
    name: O'Sassy
    url: https://osaasy.dev/
  description: |
    Server agent API for handling deployments, updates, and health checks.
    This agent is installed on servers/VPS instances and communicates with the Mithlond platform.
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticating requests from the Mithlond platform
  schemas:
...
    UpdateAgentRequest:
      type: object
      required:
        - artifact_version
      properties:
        artifact_version:
          type: string
        override_artifact_source:
          type: string
    UpdateAgentResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
...
paths:
...
  /agent/update:
    post:
      operationId: updateAgent
      summary: Update agent binary
      tags:
        - Agent
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "202":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateAgentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateAgentResponse"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateAgentResponse"
...